refactor(Badge): updates styles to new configuration
